requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
google-generativeai>=0.8.0
gunicorn==23.0.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.7.0
pywebpush>=2.0.3
